+++ lib/commands/generate_from_args.dart
@@ -3,6 +3,7 @@ import 'dart:convert';
 import 'dart:io';
 
 import 'package:args/command_runner.dart';
+import 'package:figma_vars_to_dart/services/logger/logger.dart';
 import 'package:figma_vars_to_dart/services/parser/code_entities.dart';
 
 import '../services/services.dart';
@@ -12,18 +13,47 @@ class GenerateFromArgsCommand extends Command {
   final ParserService parser;
   final WriterService writer;
   final CodeGeneratorService generator;
+  final LoggerService logger;
 
   GenerateFromArgsCommand({
     required this.figmaApi,
     required this.parser,
     required this.writer,
     required this.generator,
+    required this.logger,
   }) {
-    argParser.addOption('fileId', mandatory: true);
-    argParser.addOption('token', mandatory: true);
-    argParser.addOption('jsonOutputFile', mandatory: false);
-    argParser.addOption('dartOutputFolder',
-        mandatory: false, defaultsTo: 'lib/gen');
+    argParser.addOption(
+      'fileId',
+      mandatory: true,
+      help: 'The id of the figma file, can be found in the URL.',
+    );
+    argParser.addOption(
+      'token',
+      mandatory: true,
+      help: 'Your Figma personal access token.',
+    );
+    argParser.addOption(
+      'jsonOutputFile',
+      mandatory: false,
+      help:
+          'Optional filepath to the raw response from figma, mostly used for debugging.',
+    );
+    argParser.addOption(
+      'dartOutputFolder',
+      mandatory: false,
+      defaultsTo: 'lib/gen',
+      help:
+          'Folder location for the generated Dart files, defaults to lib/gen.',
+    );
+    argParser.addMultiOption(
+      'collectionOverrides',
+      defaultsTo: [],
+      help: 'Optional comma-separated overrides.',
+    );
+    argParser.addMultiOption(
+      'variableOverrides',
+      defaultsTo: [],
+    );
   }
 
   @override
@@ -44,10 +74,15 @@ class GenerateFromArgsCommand extends Command {
     final jsonOutput = args['jsonOutputFile'] as String?;
     final dartOutputFolder = args['dartOutputFolder'] as String;
 
-    print('Fetching the file $fileId from Figma');
+    final collectionOverrides = args['collectionOverrides'] as List<String>;
+    final variableOverrides = args['variableOverrides'] as List<String>;
+
+    logger.print('Fetching the file $fileId from Figma');
     final (response, rawResponse) = await figmaApi.getVariables(
       token: token,
       fileId: fileId,
+      variableOverrides: parseOverrides(variableOverrides),
+      collectionOverrides: parseOverrides(collectionOverrides),
     );
 
     if (jsonOutput != null) {
@@ -59,18 +94,31 @@ class GenerateFromArgsCommand extends Command {
       ]);
     }
 
-    print('Parsing json response');
+    logger.print('Parsing json response');
     final classes = parser.parse(response);
 
-    print('Converting variables to Dart');
+    logger.print('Converting variables to Dart');
     await write(dartOutputFolder, classes);
+  }
+
+  Map<String, String> parseOverrides(List<String> input) {
+    var separator = '=';
+    return {
+      for (final assignment in input)
+        if (assignment.contains(separator))
+          assignment.split(separator)[0]: assignment.split(separator)[1],
+    };
   }
 
   Future<void> write(
     String dartOutputFolder,
     List<CodeClass> classes,
   ) async {
-    final parentDirectory = dartOutputFolder.split(Platform.pathSeparator).last;
+    final parentDirectory = dartOutputFolder
+        .split(Platform.pathSeparator)
+        .where((folder) => folder.isNotEmpty)
+        .last;
+
     final files = generator.generateFiles(
       classes,
       parentDirectory,