+++ CHANGELOG.md
@@ -3,3 +3,7 @@
 
 ## 0.0.2
 - fix broken images in readme
+
+## 0.0.3
+- fix the error when the VARIABLE_ALIAS value can't be found
+- add `collectionOverrides` and `variableOverrides` parameters